Create a generic ConditionalLogger. Base SensitiveLogger on that
