+++ leaf_common/logging/sensitive_logger.py
@@ -19,10 +19,11 @@ See class comment for details.
 """
 
 from logging import Logger
-from os import getenv
+
+from leaf_common.logging.conditional_logger import ConditionalLogger
 
 
-class SensitiveLogger:
+class SensitiveLogger(ConditionalLogger):
     """
     Wraps a logger to mask sensitive information.
     Uses the same general Logger interface as the whole world already uses.
@@ -39,105 +40,4 @@ class SensitiveLogger:
 
         :param logger: The wrapped logger to redirect write() calls to.
         """
-        self.logger: Logger = logger
-        self._should_log: bool = getenv('LEAF_LOG_SENSITIVE', 'true').lower() == 'true'
-
-    def should_log(self) -> bool:
-        """
-        :return: True if this logger should log
-        """
-        return self._should_log
-
-    def debug(self, msg, *args, **kwargs):
-        """
-        Log 'msg % args' with severity 'DEBUG'.
-
-        To pass exception information, use the keyword argument exc_info with
-        a true value, e.g.
-
-        logger.debug("Houston, we have a %s", "thorny problem", exc_info=True)
-        """
-        if not self.should_log():
-            return
-        self.logger.debug(msg, *args, **kwargs)
-
-    def info(self, msg, *args, **kwargs):
-        """
-        Log 'msg % args' with severity 'INFO'.
-
-        To pass exception information, use the keyword argument exc_info with
-        a true value, e.g.
-
-        logger.info("Houston, we have a %s", "notable problem", exc_info=True)
-        """
-        if not self.should_log():
-            return
-        self.logger.info(msg, *args, **kwargs)
-
-    def warning(self, msg, *args, **kwargs):
-        """
-        Log 'msg % args' with severity 'WARNING'.
-
-        To pass exception information, use the keyword argument exc_info with
-        a true value, e.g.
-
-        logger.warning("Houston, we have a %s", "bit of a problem", exc_info=True)
-        """
-        if not self.should_log():
-            return
-        self.logger.warning(msg, *args, **kwargs)
-
-    def error(self, msg, *args, **kwargs):
-        """
-        Log 'msg % args' with severity 'ERROR'.
-
-        To pass exception information, use the keyword argument exc_info with
-        a true value, e.g.
-
-        logger.error("Houston, we have a %s", "major problem", exc_info=True)
-        """
-        if not self.should_log():
-            return
-        self.logger.error(msg, *args, **kwargs)
-
-    def exception(self, msg, *args, exc_info=True, **kwargs):
-        """
-        Convenience method for logging an ERROR with exception information.
-        """
-        if not self.should_log():
-            return
-        self.logger.exception(msg, *args, exc_info=exc_info, **kwargs)
-
-    def critical(self, msg, *args, **kwargs):
-        """
-        Log 'msg % args' with severity 'CRITICAL'.
-
-        To pass exception information, use the keyword argument exc_info with
-        a true value, e.g.
-
-        logger.critical("Houston, we have a %s", "major disaster", exc_info=True)
-        """
-        if not self.should_log():
-            return
-        self.logger.critical(msg, *args, **kwargs)
-
-    def fatal(self, msg, *args, **kwargs):
-        """
-        Don't use this method, use critical() instead.
-        """
-        if not self.should_log():
-            return
-        self.logger.fatal(msg, *args, **kwargs)
-
-    def log(self, level, msg, *args, **kwargs):
-        """
-        Log 'msg % args' with the integer severity 'level'.
-
-        To pass exception information, use the keyword argument exc_info with
-        a true value, e.g.
-
-        logger.log(level, "We have a %s", "mysterious problem", exc_info=True)
-        """
-        if not self.should_log():
-            return
-        self.logger.log(level, msg, *args, **kwargs)
+        super().__init__(logger, "LEAF_LOG_SENSITIVE")
